usb_deployment/Bybit_1h/requirements.txt
ccxt
pandas
numpy
xgboost
lightgbm
joblib
schedule
python-dotenv
